@@ -1,13 +1,13 @@
 name: orgro
 description: An Org Mode file viewer
-version: 1.55.1+159
+version: 1.66.3+197
 homepage: https://orgro.org
 
 publish_to: none
 
 environment:
-  sdk: ">=3.7.0 <4.0.0"
-  flutter: ^3.29.0
+  sdk: ">=3.9.0 <4.0.0"
+  flutter: ^3.35.7
 
 dependencies:
   cached_network_image: ^3.0.0
@@ -26,7 +26,7 @@ dependencies:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  flutter_svg: ^2.0.5
+  flutter_svg: ^2.2.0
   flutter_cache_manager: ^3.3.0
   flutter_charset_detector: ^5.0.0
   # flutter_charset_detector:
@@ -34,23 +34,29 @@ dependencies:
   flutter_slidable: ^4.0.0
   google_fonts: ^6.0.0
   http: ^1.1.0
-  intl: ^0.19.0
-  org_flutter: ^9.1.3
+  intl: ^0.20.2
+  org_flutter: ^10.0.2
   # org_flutter:
     # path: ../org_flutter
   path_provider: ^2.0.9
-  share_plus: ^10.0.1
+  share_plus: ^12.0.0
   shared_preferences: ^2.0.0
   url_launcher: ^6.0.0
   petit_bibtex: ^6.1.2
   # petit_bibtex:
     # path: ../petit_bibtex
-  open_file:
-    ^3.5.10
-  openpgp: ^3.10.0
+  open_file: ^3.5.10
+  # TODO(aaron): Unpeg pending resolution to
+  # https://github.com/jerson/flutter-openpgp/issues/83
+  openpgp: 3.10.3
+  quick_actions: ^1.1.0
+  flutter_local_notifications: ^19.4.2
+  flutter_timezone: ^5.0.0
+  timezone: ^0.10.1
+  workmanager: ^0.9.0+3
 
 dev_dependencies:
-  flutter_lints: ^5.0.0
+  flutter_lints: ^6.0.0
   flutter_test:
     sdk: flutter
 
