@@ -24,10 +24,24 @@ class AppLocalizationsUk extends AppLocalizations {
   String get menuItemOrgroManual => 'Посібник Orgro';
 
   @override
+  String get menuItemOrgManual => 'Посібник Org';
+
+  @override
   String get menuItemTestFile => 'Test File';
 
   @override
   String get menuItemAbout => 'Про застосунок';
+
+  @override
+  String get menuItemOpenUrl => 'Open URL';
+
+  @override
+  String get quickActionNewDocument => 'New Document';
+
+  @override
+  String quickActionTopPin(String name) {
+    return 'Open $name';
+  }
 
   @override
   String get appearanceModeAutomatic => 'Автоматичний';
@@ -45,7 +59,23 @@ class AppLocalizationsUk extends AppLocalizations {
   String get settingsSectionDefaultText => 'Default text';
 
   @override
+  String get settingsSectionNotifications => 'Notifications';
+
+  @override
   String get settingsSectionDataManagement => 'Data management';
+
+  @override
+  String get agendaNotificationsChannelName => 'Agenda Notifications';
+
+  @override
+  String get agendaNotificationsChannelDescription =>
+      'Notifications for Org Agenda items';
+
+  @override
+  String get agendaNotificationsActionView => 'View';
+
+  @override
+  String get settingsItemLoading => 'Loading...';
 
   @override
   String get settingsItemAppearance => 'Appearance';
@@ -57,16 +87,50 @@ class AppLocalizationsUk extends AppLocalizations {
   String get settingsItemFontFamily => 'Font';
 
   @override
+  String settingsItemInspectNotifications(num n) {
+    final intl.NumberFormat nNumberFormat = intl.NumberFormat.decimalPattern(
+      localeName,
+    );
+    final String nString = nNumberFormat.format(n);
+
+    String _temp0 = intl.Intl.pluralLogic(
+      n,
+      locale: localeName,
+      other: '$nString pending notifications',
+      one: '1 pending notification',
+      zero: 'No notifications',
+    );
+    return '$_temp0';
+  }
+
+  @override
+  String get settingsDialogNotificationsTitle => 'Pending notifications';
+
+  @override
+  String get settingsItemClearNotifications => 'Clear all notifications';
+
+  @override
   String get settingsActionClearCache => 'Clear cache';
 
   @override
   String get settingsActionResetPreferences => 'Restore defaults';
 
   @override
+  String get settingsActionResetDirectoryPermissions =>
+      'Reset directory permissions';
+
+  @override
+  String get snackbarMessageNotificationsCleared => 'Notifications cleared';
+
+  @override
   String get snackbarMessageCacheCleared => 'Кеш очищено';
 
   @override
   String get snackbarMessagePreferencesReset => 'Preferences reset';
+
+  @override
+  String get snackbarMessageDirectoryPermissionsReset =>
+      'Directory permissions reset';
 
   @override
   String get confirmResetPreferencesDialogTitle => 'Restore defaults?';
@@ -100,6 +164,9 @@ class AppLocalizationsUk extends AppLocalizations {
   String get buttonOpenOrgManual => 'Відкрити Посібник Org';
 
   @override
+  String get buttonOpenUrl => 'Open URL';
+
+  @override
   String get buttonSupport => 'Підтримка · Зворотній звʼязок';
 
   @override
@@ -108,7 +175,28 @@ class AppLocalizationsUk extends AppLocalizations {
   }
 
   @override
+  String get sectionHeaderPinnedFiles => 'Pinned files';
+
+  @override
   String get sectionHeaderRecentFiles => 'Останні файли';
+
+  @override
+  String get recentFilesSortDialogTitle => 'Sort by';
+
+  @override
+  String get sortKeyLastOpened => 'Last opened';
+
+  @override
+  String get sortKeyName => 'Name';
+
+  @override
+  String get sortKeyLocation => 'Location';
+
+  @override
+  String get sortOrderAscending => 'Ascending';
+
+  @override
+  String get sortOrderDescending => 'Descending';
 
   @override
   String get fileSourceDocuments => 'Documents';
@@ -159,11 +247,11 @@ class AppLocalizationsUk extends AppLocalizations {
 
   @override
   String searchResultSelection(num current, num total) {
-    final intl.NumberFormat currentNumberFormat = intl
-        .NumberFormat.decimalPattern(localeName);
+    final intl.NumberFormat currentNumberFormat =
+        intl.NumberFormat.decimalPattern(localeName);
     final String currentString = currentNumberFormat.format(current);
-    final intl.NumberFormat totalNumberFormat = intl
-        .NumberFormat.decimalPattern(localeName);
+    final intl.NumberFormat totalNumberFormat =
+        intl.NumberFormat.decimalPattern(localeName);
     final String totalString = totalNumberFormat.format(total);
 
     return '$currentString / $totalString';
@@ -185,6 +273,10 @@ class AppLocalizationsUk extends AppLocalizations {
 
   @override
   String get snackbarActionEnterEncryptionKey => 'Enter password';
+
+  @override
+  String get snackbarMessageNotificationPermissionsDenied =>
+      'No permission for notifications';
 
   @override
   String get dialogTitleError => 'Помилка';
@@ -273,6 +365,19 @@ class AppLocalizationsUk extends AppLocalizations {
   String get bannerBodyActionDecryptNever => 'Never';
 
   @override
+  String get bannerBodyAgendaNotifications =>
+      'Get notifications for agenda items in this file?';
+
+  @override
+  String get bannerBodyActionAgendaEnable => 'Notify me';
+
+  @override
+  String get bannerBodyActionAgendaNotNow => 'Not now';
+
+  @override
+  String get bannerBodyActionAgendaNever => 'Never';
+
+  @override
   String get saveChangesDialogTitle => 'Save changes?';
 
   @override
@@ -284,6 +389,9 @@ class AppLocalizationsUk extends AppLocalizations {
 
   @override
   String get saveActionDiscard => 'Discard';
+
+  @override
+  String get saveActionSaveAs => 'Save as';
 
   @override
   String get discardChangesDialogTitle => 'Discard changes?';
@@ -316,6 +424,12 @@ class AppLocalizationsUk extends AppLocalizations {
   String get inputCustomFilterDialogHistoryButton => 'History';
 
   @override
+  String get loadingProgressDialogTitle => 'Loading…';
+
+  @override
+  String get preparingProgressDialogTitle => 'Preparing…';
+
+  @override
   String get decryptingProgressDialogTitle => 'Decrypting…';
 
   @override
@@ -338,6 +452,12 @@ class AppLocalizationsUk extends AppLocalizations {
 
   @override
   String get createFileDefaultName => 'untitled.org';
+
+  @override
+  String get saveAsDialogTitle => 'Save as';
+
+  @override
+  String get inputUrlDialogTitle => 'Enter URL';
 
   @override
   String get startTimePickerTitle => 'Select start time';
@@ -390,8 +510,8 @@ class AppLocalizationsUk extends AppLocalizations {
 
   @override
   String errorOrgTimeout(num timeout, String code) {
-    final intl.NumberFormat timeoutNumberFormat = intl
-        .NumberFormat.decimalPattern(localeName);
+    final intl.NumberFormat timeoutNumberFormat =
+        intl.NumberFormat.decimalPattern(localeName);
     final String timeoutString = timeoutNumberFormat.format(timeout);
 
     return 'Evaluation of the below code timed out after ${timeoutString}ms.\n\n$code';
@@ -413,6 +533,16 @@ class AppLocalizationsUk extends AppLocalizations {
   @override
   String errorExternalIdNotFound(String id) {
     return 'File with ID “$id” not found';
+  }
+
+  @override
+  String errorSectionNotFound(String section) {
+    return 'Section not found: $section';
+  }
+
+  @override
+  String errorUnsupportedSearchOption(String option) {
+    return 'Unsupported search option: $option';
   }
 
   @override
