@@ -2,13 +2,20 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:orgro/l10n/app_localizations.dart';
+import 'package:orgro/src/agenda.dart';
+import 'package:orgro/src/background_tasks.dart';
 import 'package:orgro/src/cache.dart';
+import 'package:orgro/src/components/remembered_files.dart';
 import 'package:orgro/src/debug.dart';
 import 'package:orgro/src/pages/pages.dart';
 import 'package:orgro/src/preferences.dart';
+import 'package:orgro/src/quick_actions.dart';
+import 'package:orgro/src/routes/routes.dart';
 import 'package:orgro/theme.dart';
 
-void main() {
+final startKey = GlobalKey<StartPageState>();
+
+void main() async {
   LicenseRegistry.addLicense(() async* {
     yield LicenseEntryWithLineBreaks([
       'google_fonts',
@@ -25,11 +32,39 @@ void main() {
 
   runApp(buildApp());
 
-  clearTemporaryAttachments();
+  try {
+    await clearTemporaryAttachments();
+  } catch (e, s) {
+    logError(e, s);
+  }
+
+  try {
+    await cleanNotficationsLockfile();
+  } catch (e, s) {
+    logError(e, s);
+  }
 }
 
-Widget buildApp() =>
-    const SharedPreferencesProvider(waiting: _Splash(), child: _MyApp());
+Widget buildApp({bool isTest = false}) => ExecutionMode(
+  isTest: isTest,
+  child: Preferences(
+    isTest: isTest,
+    child: RememberedFiles(child: const _MyApp()),
+  ),
+);
+
+class ExecutionMode extends InheritedWidget {
+  static ExecutionMode of(BuildContext context) =>
+      context.dependOnInheritedWidgetOfExactType<ExecutionMode>()!;
+
+  const ExecutionMode({required this.isTest, super.key, required super.child});
+
+  final bool isTest;
+
+  @override
+  bool updateShouldNotify(ExecutionMode oldWidget) =>
+      isTest != oldWidget.isTest;
+}
 
 // Not the "real" splash screen; just something to cover the blank while waiting
 // for Preferences to load
@@ -45,24 +80,31 @@ class _Splash extends StatelessWidget {
   }
 }
 
-class _MyApp extends StatefulWidget {
+class _MyApp extends StatelessWidget {
   const _MyApp();
 
   @override
-  State createState() => _MyAppState();
-}
-
-class _MyAppState extends State<_MyApp> {
-  @override
-  Widget build(BuildContext context) => MaterialApp(
-    restorationScopeId: 'orgro_root',
-    localizationsDelegates: AppLocalizations.localizationsDelegates,
-    supportedLocales: AppLocalizations.supportedLocales,
-    debugShowCheckedModeBanner: !kScreenshotMode,
-    onGenerateTitle: (context) => AppLocalizations.of(context)!.appTitle,
-    theme: orgroLightTheme,
-    darkTheme: orgroDarkTheme,
-    themeMode: Preferences.of(context, PrefsAspect.appearance).themeMode,
-    home: const StartPage(),
-  );
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      restorationScopeId: 'orgro_root',
+      localizationsDelegates: AppLocalizations.localizationsDelegates,
+      supportedLocales: AppLocalizations.supportedLocales,
+      debugShowCheckedModeBanner: !kScreenshotMode,
+      onGenerateTitle: (context) => AppLocalizations.of(context)!.appTitle,
+      theme: orgroLightTheme,
+      darkTheme: orgroDarkTheme,
+      themeMode: Preferences.of(context, PrefsAspect.appearance).themeMode,
+      home: AnimatedSwitcher(
+        duration: const Duration(milliseconds: 300),
+        child: Preferences.of(context, PrefsAspect.init).isInitialized
+            ? QuickActions(
+                child: BackgroundTasks(child: StartPage(key: startKey)),
+              )
+            : const _Splash(),
+        transitionBuilder: (child, animation) =>
+            FadeTransition(opacity: animation, child: child),
+      ),
+      onGenerateRoute: onGenerateRoute,
+    );
+  }
 }
