@@ -1,6 +1,23 @@
 import 'package:flutter/material.dart';
 import 'package:orgro/l10n/app_localizations.dart';
 
+class ScrollToDocumentBoundaryAction
+    extends ContextAction<ScrollToDocumentBoundaryIntent> {
+  @override
+  void invoke(
+    covariant ScrollToDocumentBoundaryIntent intent, [
+    BuildContext? context,
+  ]) {
+    if (context != null) {
+      if (intent.forward) {
+        _scrollToBottom(context);
+      } else {
+        _scrollToTop(context);
+      }
+    }
+  }
+}
+
 void _scrollTo(ScrollController controller, double position) =>
     controller.animateTo(
       position,
@@ -25,14 +42,20 @@ class ScrollTopButton extends StatelessWidget {
   Widget build(BuildContext context) {
     return IconButton(
       icon: const Icon(Icons.keyboard_arrow_up),
-      onPressed: () => _scrollToTop(context),
+      onPressed: Actions.handler(
+        context,
+        const ScrollToDocumentBoundaryIntent(forward: false),
+      ),
     );
   }
 }
 
 PopupMenuItem<VoidCallback> scrollTopMenuItem(BuildContext context) {
   return PopupMenuItem<VoidCallback>(
-    value: () => _scrollToTop(context),
+    value: Actions.handler(
+      context,
+      ScrollToDocumentBoundaryIntent(forward: false),
+    ),
     child: Text(AppLocalizations.of(context)!.menuItemScrollTop),
   );
 }
@@ -44,14 +67,20 @@ class ScrollBottomButton extends StatelessWidget {
   Widget build(BuildContext context) {
     return IconButton(
       icon: const Icon(Icons.keyboard_arrow_down),
-      onPressed: () => _scrollToBottom(context),
+      onPressed: Actions.handler(
+        context,
+        const ScrollToDocumentBoundaryIntent(forward: true),
+      ),
     );
   }
 }
 
 PopupMenuItem<VoidCallback> scrollBottomMenuItem(BuildContext context) {
   return PopupMenuItem<VoidCallback>(
-    value: () => _scrollToBottom(context),
+    value: Actions.handler(
+      context,
+      ScrollToDocumentBoundaryIntent(forward: true),
+    ),
     child: Text(AppLocalizations.of(context)!.menuItemScrollBottom),
   );
 }
