@@ -11,21 +11,27 @@ class MySearchDelegate {
     required this.onQueryChanged,
     required this.onQuerySubmitted,
     required this.onFilterChanged,
-    String? initialQuery,
+    SearchQuery? initialQuery,
     FilterData? initialFilter,
-  }) : _searchController = TextEditingController(text: initialQuery),
+  }) : _searchController = TextEditingController(
+         text: initialQuery?.queryString,
+       ),
        _selectedFilter = ValueNotifier(initialFilter ?? FilterData.defaults()) {
     _searchController.addListener(
       debounce(
-        () => onQueryChanged(_searchController.text),
+        () => onQueryChanged(
+          SearchQuery(_searchController.text, _queryType.value),
+        ),
         const Duration(milliseconds: 500),
       ),
     );
     _selectedFilter.addListener(() => onFilterChanged(_selectedFilter.value));
+    _queryType.value = initialQuery?.type ?? QueryType.plain;
+    _queryType.addListener(() => onQuerySubmitted(query));
   }
 
-  final void Function(String) onQueryChanged;
-  final void Function(String) onQuerySubmitted;
+  final void Function(SearchQuery) onQueryChanged;
+  final void Function(SearchQuery) onQuerySubmitted;
   final void Function(FilterData) onFilterChanged;
   List<String> keywords = [];
   List<String> tags = [];
@@ -33,11 +39,18 @@ class MySearchDelegate {
   List<OrgTodoStates> todoSettings = [];
   final ValueNotifier<FilterData> _selectedFilter;
   final ValueNotifier<bool> searchMode = ValueNotifier(false);
+  final ValueNotifier<QueryType> _queryType = ValueNotifier(QueryType.plain);
   final TextEditingController _searchController;
   final FocusNode _searchFocusNode = FocusNode();
+  final _searchResultsNavigationKey =
+      GlobalKey<_SearchResultsNavigationState>();
 
-  set query(String value) {
-    _searchController.text = value;
+  SearchQuery get query =>
+      SearchQuery(_searchController.text, _queryType.value);
+
+  set query(SearchQuery value) {
+    _searchController.text = value.queryString;
+    _queryType.value = value.type;
     // Apply immediately because we debounce in the listener, and short queries
     // are swallowed, etc.
     onQuerySubmitted(value);
@@ -50,8 +63,10 @@ class MySearchDelegate {
     filterData: _selectedFilter,
     focusNode: _searchFocusNode,
     todoSettings: todoSettings,
+    queryType: _queryType,
     onClear: _clearSearchQuery,
-    onSubmitted: onQuerySubmitted,
+    onSubmitted: (value) =>
+        onQuerySubmitted(SearchQuery(value, _queryType.value)),
   );
 
   Widget buildBottomSheet(BuildContext context) {
@@ -65,6 +80,9 @@ class MySearchDelegate {
       ),
     );
   }
+
+  Widget buildSearchResultsNavigation() =>
+      _SearchResultsNavigation(key: _searchResultsNavigationKey);
 
   void dispose() {
     _searchController.dispose();
@@ -88,7 +106,7 @@ class MySearchDelegate {
   void _clearSearchQuery() {
     _searchController.clear();
     // Apply immediately because we debounce in the listener
-    onQuerySubmitted(_searchController.text);
+    onQuerySubmitted(SearchQuery('', _queryType.value));
     // It's somehow surprising to clear the filter here as well, so don't
   }
 
@@ -96,7 +114,15 @@ class MySearchDelegate {
       _searchController.value.text.isNotEmpty ||
       _selectedFilter.value.isNotEmpty;
 
-  String get queryString => _searchController.value.text;
+  void navigateSearchHits({required bool forward}) {
+    // TODO(aaron): I don't like having to use a global key to do this. It seems
+    // like it should be doable with just actions and intents. But I couldn't
+    // find a way to get Shortcuts in DocumentPage to see the Actions in
+    // SearchResultsNavigation so I hacked this up.
+    _searchResultsNavigationKey.currentState?._scrollToRelativeIndex(
+      forward ? 1 : -1,
+    );
+  }
 }
 
 class SearchField extends StatelessWidget {
@@ -105,6 +131,7 @@ class SearchField extends StatelessWidget {
     required this.filterData,
     required this.focusNode,
     required this.todoSettings,
+    required this.queryType,
     this.onClear,
     this.onSubmitted,
     super.key,
@@ -113,8 +140,23 @@ class SearchField extends StatelessWidget {
   final FocusNode focusNode;
   final List<OrgTodoStates> todoSettings;
   final ValueNotifier<FilterData> filterData;
+  final ValueNotifier<QueryType> queryType;
   final VoidCallback? onClear;
   final void Function(String)? onSubmitted;
+
+  bool _isError(String text) {
+    switch (queryType.value) {
+      case QueryType.regex:
+        try {
+          RegExp(text);
+          return false;
+        } catch (e) {
+          return true;
+        }
+      default:
+        return false;
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -130,71 +172,94 @@ class SearchField extends StatelessWidget {
         children: [
           Expanded(
             child: LayoutBuilder(
-              builder:
-                  (context, constraints) => SingleChildScrollView(
-                    scrollDirection: Axis.horizontal,
-                    child: ValueListenableBuilder(
-                      valueListenable: filterData,
-                      builder:
-                          (context, filter, _) => Row(
-                            children: [
-                              ...[
-                                SelectedFilterChips(
-                                  filter: filter,
-                                  todoSettings: todoSettings,
-                                  onChange: (value) => filterData.value = value,
-                                ),
-                                if (filter.isNotEmpty)
-                                  IconTheme.merge(
-                                    data: iconTheme,
-                                    child: const Icon(Icons.drag_indicator),
-                                  ),
-                              ].separatedBy(const SizedBox(width: 8)),
-                              ConstrainedBox(
-                                constraints:
-                                    filter.isNotEmpty
-                                        ? BoxConstraints.tightFor(
-                                          width:
-                                              constraints.maxWidth -
-                                              IconTheme.of(context).size!,
+              builder: (context, constraints) => SingleChildScrollView(
+                scrollDirection: Axis.horizontal,
+                child: ValueListenableBuilder(
+                  valueListenable: filterData,
+                  builder: (context, filter, _) => Row(
+                    children: [
+                      ...[
+                        SelectedFilterChips(
+                          filter: filter,
+                          todoSettings: todoSettings,
+                          onChange: (value) => filterData.value = value,
+                        ),
+                        if (filter.isNotEmpty)
+                          IconTheme.merge(
+                            data: iconTheme,
+                            child: const Icon(Icons.drag_indicator),
+                          ),
+                      ].separatedBy(const SizedBox(width: 8)),
+                      ConstrainedBox(
+                        constraints: filter.isNotEmpty
+                            ? BoxConstraints.tightFor(
+                                width:
+                                    constraints.maxWidth -
+                                    IconTheme.of(context).size!,
+                              )
+                            : constraints,
+                        child: ValueListenableBuilder<TextEditingValue>(
+                          valueListenable: _controller,
+                          builder: (context, value, child) {
+                            final isError = _isError(value.text);
+                            return TextField(
+                              autofocus: true,
+                              focusNode: focusNode,
+                              style: style,
+                              controller: _controller,
+                              textInputAction: TextInputAction.search,
+                              cursorColor: theme.colorScheme.secondary,
+                              onSubmitted: onSubmitted,
+                              decoration: InputDecoration(
+                                hintText: AppLocalizations.of(
+                                  context,
+                                )!.hintTextSearch,
+                                border: InputBorder.none,
+                                prefixIcon: IconTheme.merge(
+                                  data: isError
+                                      ? IconThemeData(
+                                          color: theme.colorScheme.error,
                                         )
-                                        : constraints,
-                                child: TextField(
-                                  autofocus: true,
-                                  focusNode: focusNode,
-                                  style: style,
-                                  controller: _controller,
-                                  textInputAction: TextInputAction.search,
-                                  cursorColor: theme.colorScheme.secondary,
-                                  onSubmitted: onSubmitted,
-                                  decoration: InputDecoration(
-                                    hintText:
-                                        AppLocalizations.of(
-                                          context,
-                                        )!.hintTextSearch,
-                                    border: InputBorder.none,
-                                    prefixIcon: IconTheme.merge(
-                                      data: iconTheme,
-                                      child: const Icon(Icons.search),
-                                    ),
-                                  ),
+                                      : iconTheme,
+                                  child: isError
+                                      ? const Icon(Icons.error)
+                                      : const Icon(Icons.search),
                                 ),
                               ),
-                            ],
-                          ),
-                    ),
+                            );
+                          },
+                        ),
+                      ),
+                    ],
                   ),
+                ),
+              ),
+            ),
+          ),
+          ValueListenableBuilder<QueryType>(
+            valueListenable: queryType,
+            builder: (context, value, child) => TextButton(
+              onPressed: () => queryType.value = switch (value) {
+                QueryType.plain => QueryType.regex,
+                QueryType.regex => QueryType.plain,
+              },
+              style: TextButton.styleFrom(foregroundColor: color),
+              child: Text(switch (value) {
+                QueryType.plain => 'Aa',
+                QueryType.regex => '.*',
+              }),
             ),
           ),
           ValueListenableBuilder<TextEditingValue>(
             valueListenable: _controller,
-            builder:
-                (context, value, child) =>
-                    value.text.isNotEmpty ? child! : const SizedBox.shrink(),
+            builder: (context, value, child) =>
+                value.text.isNotEmpty ? child! : const SizedBox.shrink(),
             child: IconTheme.merge(
               data: iconTheme,
               child: IconButton(
-                icon: const Icon(Icons.clear),
+                // Icons.clear looks too much like Icons.close so we substitute
+                // Icons.cancel, which is the same but the X is in a circle.
+                icon: const Icon(Icons.cancel),
                 onPressed: onClear,
               ),
             ),
@@ -205,15 +270,15 @@ class SearchField extends StatelessWidget {
   }
 }
 
-class SearchResultsNavigation extends StatefulWidget {
-  const SearchResultsNavigation({super.key});
+class _SearchResultsNavigation extends StatefulWidget {
+  const _SearchResultsNavigation({super.key});
 
   @override
-  State<SearchResultsNavigation> createState() =>
+  State<_SearchResultsNavigation> createState() =>
       _SearchResultsNavigationState();
 }
 
-class _SearchResultsNavigationState extends State<SearchResultsNavigation> {
+class _SearchResultsNavigationState extends State<_SearchResultsNavigation> {
   int _i = -1;
   List<SearchResultKey> _keys = [];
   late OrgControllerData _controller;
@@ -266,8 +331,8 @@ class _SearchResultsNavigationState extends State<SearchResultsNavigation> {
                 _i == -1
                     ? AppLocalizations.of(context)!.searchHits(_keys.length)
                     : AppLocalizations.of(
-                      context,
-                    )!.searchResultSelection(_i + 1, _keys.length),
+                        context,
+                      )!.searchResultSelection(_i + 1, _keys.length),
                 textAlign: TextAlign.center,
                 style: DefaultTextStyle.of(context).style.copyWith(
                   fontFeatures: const [FontFeature.tabularFigures()],
@@ -339,8 +404,7 @@ class _DisablableMiniFloatingActionButton extends StatelessWidget {
     );
   }
 
-  Color? _backgroundColor(BuildContext context) =>
-      _enabled
-          ? null // default
-          : Theme.of(context).disabledColor;
+  Color? _backgroundColor(BuildContext context) => _enabled
+      ? null // default
+      : Theme.of(context).disabledColor;
 }
