@@ -13,21 +13,26 @@ PopupMenuEntry<VoidCallback> textScaleMenuItem(
   );
 }
 
-Widget textScaleListItem(BuildContext context) {
-  final viewSettings = ViewSettings.of(context);
-  return ListTile(
-    title: Row(
-      children: [
-        Expanded(
-          child: Text(AppLocalizations.of(context)!.settingsItemTextScale),
-        ),
-        TextSizeAdjuster(
-          value: viewSettings.textScale,
-          onChanged: (value) => viewSettings.textScale = value,
-        ),
-      ],
-    ),
-  );
+class TextScaleSettingListItem extends StatelessWidget {
+  const TextScaleSettingListItem({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final viewSettings = ViewSettings.of(context);
+    return ListTile(
+      title: Row(
+        children: [
+          Expanded(
+            child: Text(AppLocalizations.of(context)!.settingsItemTextScale),
+          ),
+          TextSizeAdjuster(
+            value: viewSettings.textScale,
+            onChanged: (value) => viewSettings.textScale = value,
+          ),
+        ],
+      ),
+    );
+  }
 }
 
 PopupMenuEntry<VoidCallback> fontFamilyMenuItem(
@@ -40,18 +45,23 @@ PopupMenuEntry<VoidCallback> fontFamilyMenuItem(
   );
 }
 
-Widget fontFamilyListItem(BuildContext context) {
-  final viewSettings = ViewSettings.of(context);
-  return ListTile(
-    title: Text(AppLocalizations.of(context)!.settingsItemFontFamily),
-    subtitle: Text(viewSettings.fontFamily),
-    onTap: () async {
-      final selection = await _chooseFont(context, viewSettings.fontFamily);
-      if (selection != null) {
-        viewSettings.fontFamily = selection;
-      }
-    },
-  );
+class FontFamilySettingListItem extends StatelessWidget {
+  const FontFamilySettingListItem({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final viewSettings = ViewSettings.of(context);
+    return ListTile(
+      title: Text(AppLocalizations.of(context)!.settingsItemFontFamily),
+      subtitle: Text(viewSettings.fontFamily),
+      onTap: () async {
+        final selection = await _chooseFont(context, viewSettings.fontFamily);
+        if (selection != null) {
+          viewSettings.fontFamily = selection;
+        }
+      },
+    );
+  }
 }
 
 class TextStyleButton extends StatefulWidget {
@@ -78,8 +88,8 @@ class _TextStyleButtonState extends State<TextStyleButton>
   Widget build(BuildContext context) {
     return IconButton(
       icon: const Icon(Icons.format_size),
-      onPressed:
-          () => Overlay.of(context).insertAll(_overlays(context).toList()),
+      onPressed: () =>
+          Overlay.of(context).insertAll(_overlays(context).toList()),
     );
   }
 
@@ -301,18 +311,18 @@ class _FontFamilySelectorState extends State<FontFamilySelector> {
 Future<String?> _chooseFont(BuildContext context, String currentValue) async =>
     showDialog<String>(
       context: context,
-      builder:
-          (context) => SimpleDialog(
-            children: [
-              for (final family in availableFontFamilies.toList(growable: false)
-                ..sort())
-                CheckboxListTile(
-                  value: currentValue == family,
-                  title: Text(family),
-                  onChanged: (_) => Navigator.pop(context, family),
-                ),
-            ],
-          ),
+      builder: (context) => SimpleDialog(
+        children: [
+          for (final family in availableFontFamilies.toList(
+            growable: false,
+          )..sort())
+            CheckboxListTile(
+              value: currentValue == family,
+              title: Text(family),
+              onChanged: (_) => Navigator.pop(context, family),
+            ),
+        ],
+      ),
     );
 
 /// A popup menu item that doesn't close when tapped and doesn't provide its own
