@@ -3,6 +3,7 @@ import 'dart:math';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:org_flutter/org_flutter.dart';
+import 'package:orgro/src/agenda.dart';
 import 'package:orgro/src/data_source.dart';
 import 'package:orgro/src/debug.dart';
 import 'package:orgro/src/encryption.dart';
@@ -43,7 +44,7 @@ class _DocumentProviderState extends State<DocumentProvider> {
   void initState() {
     super.initState();
     _docs = [widget.doc];
-    _analyses = [const DocumentAnalysis()];
+    _analyses = [DocumentAnalysis.empty()];
     _analyze(widget.doc).then((analysis) {
       setState(() => _analyses[0] = analysis);
     });
@@ -53,8 +54,10 @@ class _DocumentProviderState extends State<DocumentProvider> {
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
-    final accessibleDirs =
-        Preferences.of(context, PrefsAspect.accessibleDirs).accessibleDirs;
+    final accessibleDirs = Preferences.of(
+      context,
+      PrefsAspect.accessibleDirs,
+    ).accessibleDirs;
     _resolveDataSourceParent(accessibleDirs).then((dataSource) {
       if (dataSource != null) {
         setState(() => _dataSource = dataSource);
@@ -197,6 +200,7 @@ class DocumentAnalysis {
     var hasRemoteImages = false;
     var hasRelativeLinks = false;
     var hasEncryptedContent = false;
+    var hasAgendaEntries = false;
     doc.visit<OrgNode>((node) {
       if (node is OrgLink) {
         hasRemoteImages |=
@@ -208,6 +212,10 @@ class DocumentAnalysis {
         }
       } else if (node is OrgPgpBlock) {
         hasEncryptedContent = true;
+      } else if (node is OrgSection) {
+        // TODO(aaron): Is this right? Is it efficient?
+        // TODO(aaron): Should we detect only *pending* entries?
+        hasAgendaEntries = node.scheduledAt.isNotEmpty;
       }
       return !hasRemoteImages ||
           (!hasRelativeLinks && canResolveRelativeLinks) ||
@@ -224,10 +232,8 @@ class DocumentAnalysis {
       if (keyword != null) {
         keywords.add(keyword);
       }
-      final sectionTags = section.headline.tags;
-      if (sectionTags != null) {
-        tags.addAll(sectionTags.values);
-      }
+      final sectionTags = section.tags;
+      tags.addAll(sectionTags);
       final priority = section.headline.priority?.value;
       if (priority != null) {
         priorities.add(priority);
@@ -236,30 +242,38 @@ class DocumentAnalysis {
     });
 
     return DocumentAnalysis(
+      loaded: true,
       hasRemoteImages: hasRemoteImages,
       hasRelativeLinks: hasRelativeLinks,
       hasEncryptedContent: hasEncryptedContent,
       needsEncryption: needsEncryption,
+      hasAgendaEntries: hasAgendaEntries,
       keywords: keywords.toList(growable: false),
       tags: tags.toList(growable: false),
       priorities: priorities.toList(growable: false),
     );
   }
 
+  factory DocumentAnalysis.empty() => const DocumentAnalysis(loaded: false);
+
   const DocumentAnalysis({
+    required this.loaded,
     this.hasRemoteImages,
     this.hasRelativeLinks,
     this.hasEncryptedContent,
     this.needsEncryption,
+    this.hasAgendaEntries,
     this.keywords,
     this.tags,
     this.priorities,
   });
 
+  final bool loaded;
   final bool? hasRemoteImages;
   final bool? hasRelativeLinks;
   final bool? hasEncryptedContent;
   final bool? needsEncryption;
+  final bool? hasAgendaEntries;
   final List<String>? keywords;
   final List<String>? tags;
   final List<String>? priorities;
@@ -267,20 +281,24 @@ class DocumentAnalysis {
   @override
   bool operator ==(Object other) =>
       other is DocumentAnalysis &&
+      loaded == other.loaded &&
       hasRemoteImages == other.hasRemoteImages &&
       hasRelativeLinks == other.hasRelativeLinks &&
       hasEncryptedContent == other.hasEncryptedContent &&
       needsEncryption == other.needsEncryption &&
+      hasAgendaEntries == other.hasAgendaEntries &&
       listEquals(keywords, other.keywords) &&
       listEquals(tags, other.tags) &&
       listEquals(priorities, other.priorities);
 
   @override
   int get hashCode => Object.hash(
+    loaded,
     hasRemoteImages,
     hasRelativeLinks,
     hasEncryptedContent,
     needsEncryption,
+    hasAgendaEntries,
     keywords == null ? null : Object.hashAll(keywords!),
     tags == null ? null : Object.hashAll(tags!),
     priorities == null ? null : Object.hashAll(priorities!),
