@@ -11,32 +11,26 @@ class OrgroError implements Exception {
         return switch (obj) {
           OrgParserError(result: final result) => OrgroError(
             'Parser error',
-            localizedMessage:
-                (context) => AppLocalizations.of(
-                  context,
-                )!.errorOrgParser(result.toString()),
+            localizedMessage: (context) =>
+                AppLocalizations.of(context)!.errorOrgParser(result.toString()),
           ),
           OrgExecutionError(code: final code, cause: final cause) => OrgroError(
             'Execution error',
-            localizedMessage:
-                (context) => AppLocalizations.of(
-                  context,
-                )!.errorOrgExecution(cause.toString(), code),
+            localizedMessage: (context) => AppLocalizations.of(
+              context,
+            )!.errorOrgExecution(cause.toString(), code),
           ),
           OrgTimeoutError(code: final code, timeLimit: final limit) =>
             OrgroError(
               'Timeout error',
-              localizedMessage:
-                  (context) => AppLocalizations.of(
-                    context,
-                  )!.errorOrgTimeout(limit.inMilliseconds, code),
+              localizedMessage: (context) => AppLocalizations.of(
+                context,
+              )!.errorOrgTimeout(limit.inMilliseconds, code),
             ),
           OrgArgumentError(item: final item) => OrgroError(
             'Argument error',
-            localizedMessage:
-                (context) => AppLocalizations.of(
-                  context,
-                )!.errorOrgArgument(item.toString()),
+            localizedMessage: (context) =>
+                AppLocalizations.of(context)!.errorOrgArgument(item.toString()),
           ),
         };
       default:
