@@ -5,6 +5,7 @@ import 'package:file_picker_writable/file_picker_writable.dart';
 import 'package:flutter/material.dart';
 import 'package:orgro/l10n/app_localizations.dart';
 import 'package:orgro/src/data_source.dart';
+import 'package:orgro/src/pages/start/util.dart';
 
 Future<NativeDataSource?> pickFile() async =>
     FilePickerWritable().openFile(LoadedNativeDataSource.fromExternal);
@@ -17,6 +18,12 @@ Future<NativeDataSource?> createAndLoadFile(String fileName) async {
   return fileInfo == null ? null : readFileWithIdentifier(fileInfo.identifier);
 }
 
+Future<FileInfo?> createAndSaveFile(String fileName, String content) async =>
+    await FilePickerWritable().openFileForCreate(
+      fileName: fileName,
+      writer: (file) => file.writeAsString(content),
+    );
+
 Future<NativeDirectoryInfo?> pickDirectory({String? initialDirUri}) async {
   final dirInfo = await FilePickerWritable().openDirectory(
     initialDirUri: initialDirUri,
@@ -24,10 +31,10 @@ Future<NativeDirectoryInfo?> pickDirectory({String? initialDirUri}) async {
   return dirInfo == null
       ? null
       : NativeDirectoryInfo(
-        dirInfo.fileName ?? 'unknown',
-        dirInfo.identifier,
-        dirInfo.uri,
-      );
+          dirInfo.fileName ?? 'unknown',
+          dirInfo.identifier,
+          dirInfo.uri,
+        );
 }
 
 Future<NativeDataSource> readFileWithIdentifier(String identifier) async =>
@@ -48,10 +55,9 @@ mixin PlatformOpenHandler<T extends StatefulWidget> on State<T> {
   @override
   void initState() {
     super.initState();
-    _filePickerState =
-        FilePickerWritable().init()
-          ..registerFileOpenHandler(_loadFile)
-          ..registerErrorEventHandler(_handleError);
+    _filePickerState = FilePickerWritable().init()
+      ..registerFileOpenHandler(_loadFile)
+      ..registerErrorEventHandler(_handleError);
   }
 
   Future<bool> _loadFile(FileInfo fileInfo, File file) async {
@@ -62,10 +68,10 @@ mixin PlatformOpenHandler<T extends StatefulWidget> on State<T> {
       await _displayError(e.toString());
       return false;
     }
-    return loadFileFromPlatform(openFileInfo);
+    if (!mounted) return false;
+    await loadAndRememberFile(context, openFileInfo);
+    return true;
   }
-
-  Future<bool> loadFileFromPlatform(NativeDataSource info);
 
   Future<bool> _handleError(ErrorEvent event) async {
     await _displayError(event.message);
@@ -74,11 +80,10 @@ mixin PlatformOpenHandler<T extends StatefulWidget> on State<T> {
 
   Future<void> _displayError(String message) async => showDialog<void>(
     context: context,
-    builder:
-        (context) => SimpleDialog(
-          title: Text(AppLocalizations.of(context)!.dialogTitleError),
-          children: [ListTile(title: Text(message))],
-        ),
+    builder: (context) => SimpleDialog(
+      title: Text(AppLocalizations.of(context)!.dialogTitleError),
+      children: [ListTile(title: Text(message))],
+    ),
   );
 
   @override
