@@ -18,13 +18,9 @@ extension EncryptionHandler on DocumentPageState {
     });
     final password = await showDialog<String>(
       context: context,
-      builder:
-          (context) => InputPasswordDialog(
-            title:
-                AppLocalizations.of(
-                  context,
-                )!.inputDecryptionPasswordDialogTitle,
-          ),
+      builder: (context) => InputPasswordDialog(
+        title: AppLocalizations.of(context)!.inputDecryptionPasswordDialogTitle,
+      ),
     );
     if (password == null) return;
     if (!mounted) return;
@@ -59,8 +55,8 @@ extension EncryptionHandler on DocumentPageState {
         final enclosingCryptSection =
             newDoc.findContainingTree(
               block,
-              where:
-                  (tree) => tree is OrgSection && tree.tags.contains('crypt'),
+              where: (tree) =>
+                  tree is OrgSection && tree.tags.contains('crypt'),
             ) ??
             newDoc.findContainingTree(block)!;
         if (enclosingCryptSection is OrgSection) {
