@@ -4,6 +4,7 @@ import 'package:orgro/src/attachments.dart';
 import 'package:orgro/src/components/document_provider.dart';
 import 'package:orgro/src/components/image.dart';
 import 'package:orgro/src/components/view_settings.dart';
+import 'package:orgro/src/navigation.dart';
 import 'package:orgro/src/pages/document/document.dart';
 import 'package:orgro/src/preferences.dart';
 
@@ -14,7 +15,7 @@ extension ImageHandler on DocumentPageState {
     }
     final doc = DocumentProvider.of(context).doc;
     try {
-      final fileLink = convertLinkResolvingAttachments(doc, link);
+      final fileLink = convertLinkResolvingAttachments(context, doc, link);
       if (fileLink.isRelative) {
         return _loadLocalImage(link, fileLink);
       }
@@ -31,7 +32,11 @@ extension ImageHandler on DocumentPageState {
     if (viewSettings.remoteImagesPolicy != RemoteImagesPolicy.allow) {
       return null;
     }
-    return RemoteImage(link);
+    return GestureDetector(
+      onLongPress: () =>
+          showInteractive(context, link.location, RemoteImage(link)),
+      child: RemoteImage(link, scaled: true),
+    );
   }
 
   Widget? _loadLocalImage(OrgLink link, OrgFileLink fileLink) {
@@ -39,10 +44,18 @@ extension ImageHandler on DocumentPageState {
     if (source.needsToResolveParent) {
       return null;
     }
-    return LocalImage(
-      link: link,
-      dataSource: source,
-      relativePath: fileLink.body,
+    return GestureDetector(
+      onLongPress: () => showInteractive(
+        context,
+        fileLink.body,
+        LocalImage(link: link, dataSource: source, relativePath: fileLink.body),
+      ),
+      child: LocalImage(
+        link: link,
+        dataSource: source,
+        relativePath: fileLink.body,
+        minimizeSize: true,
+      ),
     );
   }
 }
