@@ -1,4 +1,4 @@
-export 'package:orgro/src/pages/document/document.dart';
-export 'package:orgro/src/pages/error.dart';
-export 'package:orgro/src/pages/progress.dart';
-export 'package:orgro/src/pages/start.dart';
+export './document/document.dart';
+export './error.dart';
+export './progress.dart';
+export './start/start.dart';
